+++ CHANGELOG.md
@@ -7,6 +7,7 @@ ChangeLog
 
 - IMPROVEMENT: add "ase setup enable" and "ase setup disable" for enabling/disabling ASE in the agent tool
 - IMPROVEMENT: truncate IMPLEMENTATION DRAFT section in "ase-task-edit" skill
+- IMPROVEMENT: add a ase-getopt.md (plugin) and ase-getopt.ts (tool) for option parsing
 - BUGFIX: fix bundling of plugin into tool
 
 0.0.32 (2026-05-18)